move state into package
