+++ src/main/scala/votingapp/PollConflictResolutionMechanism.scala
@@ -1,7 +1,7 @@
 package votingapp
 
-import rover.rdo.AtomicObjectState
 import rover.rdo.conflict.{ConflictResolutionMechanism, ConflictedState}
+import rover.rdo.state.AtomicObjectState
 
 class PollConflictResolutionMechanism extends ConflictResolutionMechanism[Votes] {
 	override def resolveConflict(conflictedState: ConflictedState[Votes]): AtomicObjectState[Votes] = {
